SentimentAnalysis/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
langdetect==1.0.9
lxml==5.2.2
MarkupSafe==2.1.5
multidict==6.0.5
playwright==1.45.1
pyahocorasick==2.1.0
pyee==11.1.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
youtube-transcript-api==0.6.2
